Extract ScrollTop logic, consolidate shared UI, relocate feature UI, closes #24
