+++ src/app/learn/page.tsx
@@ -2,8 +2,8 @@
 
 import { useState, Suspense } from 'react'
 import { useSearchParams, useRouter } from 'next/navigation'
-import Dashboard from '@/app/ui/learn/Dashboard'
-import Project from '@/app/ui/learn/Project'
+import Dashboard from '@/components/learn/Dashboard'
+import Project from '@/components/learn/Project'
 import { projectData } from '@/data/project-modules'
 
 function Page() {
