+++ src/components/ide/CodeEditor.tsx
